Cleanup things a bit

+++ WordPress/WordPressTest/ReaderTopicServiceTest.swift
@@ -3,9 +3,10 @@ import Foundation
 import CoreData
 import WordPress
 
-class ReaderTopicSwiftTest: XCTestCase {
+final class ReaderTopicSwiftTest: XCTestCase {
 
-    var testContextManager: TestContextManager?
+    private var testContextManager: TestContextManager?
+    private var context: NSManagedObjectContext?
     let expectationTimeout = 5.0
 
     // MARK: - Lifecycle
@@ -13,9 +14,11 @@ class ReaderTopicSwiftTest: XCTestCase {
     override func setUp() {
         super.setUp()
         testContextManager = TestContextManager.sharedInstance()
+        context = testContextManager?.mainContext
     }
 
     override func tearDown() {
+        context = nil
         TestContextManager.overrideSharedInstance(nil)
         testContextManager = nil
         super.tearDown()
@@ -24,59 +27,68 @@ class ReaderTopicSwiftTest: XCTestCase {
     // MARK: - Config / Helpers
 
     func seedTopics() {
-        let context = testContextManager?.mainContext
+        guard let context = context else {
+            XCTFail("Context is nil")
+            return
+        }
 
-        let topic1 = NSEntityDescription.insertNewObject(forEntityName: ReaderListTopic.classNameWithoutNamespaces(), into: context!) as! ReaderListTopic
+        let topic1 = NSEntityDescription.insertNewObject(forEntityName: ReaderListTopic.classNameWithoutNamespaces(), into: context) as! ReaderListTopic
         topic1.path = "/list/topic1"
         topic1.title = "topic1"
         topic1.type = ReaderListTopic.TopicType
 
-        let topic2 = NSEntityDescription.insertNewObject(forEntityName: ReaderTagTopic.classNameWithoutNamespaces(), into: context!) as! ReaderTagTopic
+        let topic2 = NSEntityDescription.insertNewObject(forEntityName: ReaderTagTopic.classNameWithoutNamespaces(), into: context) as! ReaderTagTopic
         topic2.path = "/tags/topic2"
         topic2.title = "topic2"
         topic2.type = ReaderTagTopic.TopicType
 
-        let topic3 = NSEntityDescription.insertNewObject(forEntityName: ReaderTagTopic.classNameWithoutNamespaces(), into: context!) as! ReaderTagTopic
+        let topic3 = NSEntityDescription.insertNewObject(forEntityName: ReaderTagTopic.classNameWithoutNamespaces(), into: context) as! ReaderTagTopic
         topic3.path = "/list/topic3"
         topic3.title = "topic3"
         topic3.type = ReaderTagTopic.TopicType
 
         do {
-            try context?.save()
+            try context.save()
         } catch let error as NSError {
             XCTAssertNil(error, "Error seeding topics")
         }
     }
 
     func countTopics() -> Int {
-        let context = testContextManager?.mainContext
+        guard let context = context else {
+            XCTFail("Context is nil")
+            return 0
+        }
         let request = NSFetchRequest<NSFetchRequestResult>(entityName: ReaderAbstractTopic.classNameWithoutNamespaces())
-        return try! context!.count(for: request)
+        return try! context.count(for: request)
     }
 
     func seedPostsForTopic(_ topic: ReaderAbstractTopic) {
-        let context = testContextManager?.mainContext
+        guard let context = context else {
+            XCTFail("Context is nil")
+            return
+        }
 
-        let post1 = NSEntityDescription.insertNewObject(forEntityName: ReaderPost.classNameWithoutNamespaces(), into: context!) as! ReaderPost
+        let post1 = NSEntityDescription.insertNewObject(forEntityName: ReaderPost.classNameWithoutNamespaces(), into: context) as! ReaderPost
         post1.postID = NSNumber(value: 1)
         post1.postTitle = "post1"
         post1.content = "post1"
         post1.topic = topic
 
-        let post2 = NSEntityDescription.insertNewObject(forEntityName: ReaderPost.classNameWithoutNamespaces(), into: context!) as! ReaderPost
+        let post2 = NSEntityDescription.insertNewObject(forEntityName: ReaderPost.classNameWithoutNamespaces(), into: context) as! ReaderPost
         post2.postID = NSNumber(value: 2)
         post2.postTitle = "post2"
         post2.content = "post2"
         post2.topic = topic
 
-        let post3 = NSEntityDescription.insertNewObject(forEntityName: ReaderPost.classNameWithoutNamespaces(), into: context!) as! ReaderPost
+        let post3 = NSEntityDescription.insertNewObject(forEntityName: ReaderPost.classNameWithoutNamespaces(), into: context) as! ReaderPost
         post3.postID = NSNumber(value: 3)
         post3.postTitle = "post3"
         post3.content = "post3"
         post3.topic = topic
 
         do {
-            try context?.save()
+            try context.save()
         } catch let error as NSError {
             XCTAssertNil(error, "Error seeding posts")
         }
@@ -130,12 +142,15 @@ class ReaderTopicSwiftTest: XCTestCase {
     results from the REST API.
     */
     func testUnsubscribedTopicIsRemovedDuringSync() {
+        guard let context = context else {
+            XCTFail("Context is nil")
+            return
+        }
         let remoteTopics = remoteTopicsForTests()
 
         // Setup
         var expect = expectation(description: "topics saved expectation")
-        let context = testContextManager?.mainContext
-        let service = ReaderTopicService(managedObjectContext: context!)
+        let service = ReaderTopicService(managedObjectContext: context)
         service.mergeMenuTopics(remoteTopics, withSuccess: { () -> Void in
             expect.fulfill()
         })
@@ -143,7 +158,7 @@ class ReaderTopicSwiftTest: XCTestCase {
 
         // Topics exist in the context
         let request = NSFetchRequest<NSFetchRequestResult>(entityName: ReaderTagTopic.classNameWithoutNamespaces())
-        var count = try! context?.count(for: request)
+        var count = try! context.count(for: request)
         XCTAssertEqual(count, remoteTopics.count, "Number of topics in context did not match expectations")
 
         // Merge new set of topics
@@ -155,11 +170,11 @@ class ReaderTopicSwiftTest: XCTestCase {
         waitForExpectations(timeout: expectationTimeout, handler: nil)
 
         // Make sure the missing topics were removed when merged
-        count = try! context?.count(for: request)
+        count = try! context.count(for: request)
         XCTAssertEqual(count, 1, "Number of topics in context did not match expectations")
         do {
-            let results = try context?.fetch(request)
-            let topic = results?.first as! ReaderTagTopic
+            let results = try context.fetch(request)
+            let topic = results.first as! ReaderTagTopic
             XCTAssertEqual(topic.tagID, foo?.topicID, "The topic returned was not the one expected.")
         } catch let error as NSError {
             XCTAssertNil(error, "Error executing fetch request.")
@@ -171,6 +186,10 @@ class ReaderTopicSwiftTest: XCTestCase {
     results from the REST API.
     */
     func testNewlySubscribedTopicIsAddedDuringSync() {
+        guard let context = context else {
+            XCTFail("Context is nil")
+            return
+        }
         var remoteTopics = remoteTopicsForTests()
         let foo = remoteTopics.first
 
@@ -178,8 +197,7 @@ class ReaderTopicSwiftTest: XCTestCase {
 
         // Setup
         var expect = expectation(description: "topics saved expectation")
-        let context = testContextManager?.mainContext
-        let service = ReaderTopicService(managedObjectContext: context!)
+        let service = ReaderTopicService(managedObjectContext: context)
         service.mergeMenuTopics(startingTopics, withSuccess: { () -> Void in
             expect.fulfill()
         })
@@ -189,7 +207,7 @@ class ReaderTopicSwiftTest: XCTestCase {
         let sortDescriptor = NSSortDescriptor(key: "tagID", ascending: true)
         let request = NSFetchRequest<NSFetchRequestResult>(entityName: ReaderTagTopic.classNameWithoutNamespaces())
         request.sortDescriptors = [sortDescriptor]
-        var count = try! context?.count(for: request)
+        var count = try! context.count(for: request)
         XCTAssertEqual(count, startingTopics.count, "Number of topics in context did not match expectations")
 
         // Merge new set of topics
@@ -200,12 +218,12 @@ class ReaderTopicSwiftTest: XCTestCase {
         waitForExpectations(timeout: expectationTimeout, handler: nil)
 
         // make sure the missing topics were added
-        count = try! context?.count(for: request)
+        count = try! context.count(for: request)
         XCTAssertEqual(count, remoteTopics.count, "Number of topics in context did not match expectations")
 
         do {
-            let results = try context?.fetch(request)
-            let topic = results?.first as! ReaderTagTopic
+            let results = try context.fetch(request)
+            let topic = results.first as! ReaderTagTopic
             XCTAssertEqual(topic.tagID, foo!.topicID, "The topic returned was not the one expected.")
         } catch let error as NSError {
             XCTAssertNil(error, "Error executing fetch request.")
@@ -216,12 +234,15 @@ class ReaderTopicSwiftTest: XCTestCase {
     Ensure that a default topic can be set and retrieved.
     */
     func testGettingSettingCurrentTopic() {
+        guard let context = context else {
+            XCTFail("Context is nil")
+            return
+        }
         let remoteTopics = remoteAndDefaultTopicForTests()
 
         // Setup
         let expect = expectation(description: "topics saved expectation")
-        let context = testContextManager?.mainContext
-        let service = ReaderTopicService(managedObjectContext: context!)
+        let service = ReaderTopicService(managedObjectContext: context)
         service.currentTopic = nil
 
         // Current topic is not nil after a sync
@@ -236,12 +257,12 @@ class ReaderTopicSwiftTest: XCTestCase {
         let request = NSFetchRequest<NSFetchRequestResult>(entityName: ReaderAbstractTopic.classNameWithoutNamespaces())
         request.sortDescriptors = [sortDescriptor]
 
-        let results = try! context?.fetch(request)
+        let results = try! context.fetch(request)
 
-        var topic = results?.last as! ReaderAbstractTopic
+        var topic = results.last as! ReaderAbstractTopic
         XCTAssertEqual(service.currentTopic.type, ReaderDefaultTopic.TopicType, "The curent topic should have been a default topic")
 
-        topic = results?.first as! ReaderAbstractTopic
+        topic = results.first as! ReaderAbstractTopic
         service.currentTopic = topic
 
         XCTAssertEqual(service.currentTopic.path, topic.path, "The current topic did not match the topic we assiged to it")
@@ -251,10 +272,13 @@ class ReaderTopicSwiftTest: XCTestCase {
     Ensure all topics are deleted when an account is changed.
     */
     func testDeleteAllTopics() {
+        guard let context = context else {
+            XCTFail("Context is nil")
+            return
+        }
         seedTopics()
         XCTAssertFalse(countTopics() == 0, "The number of seeded topics should not be zero")
-        let context = testContextManager?.mainContext
-        let service = ReaderTopicService(managedObjectContext: context!)
+        let service = ReaderTopicService(managedObjectContext: context)
         service.deleteAllTopics()
         XCTAssertTrue(countTopics() == 0, "The number of seeded topics should be zero")
     }
@@ -264,8 +288,11 @@ class ReaderTopicSwiftTest: XCTestCase {
     */
     func testPostsDeletedWhenTopicDeleted() {
         // setup
-        let context = testContextManager?.mainContext
-        let topic = NSEntityDescription.insertNewObject(forEntityName: ReaderListTopic.classNameWithoutNamespaces(), into: context!) as! ReaderListTopic
+        guard let context = context else {
+            XCTFail("Context is nil")
+            return
+        }
+        let topic = NSEntityDescription.insertNewObject(forEntityName: ReaderListTopic.classNameWithoutNamespaces(), into: context) as! ReaderListTopic
         topic.path = "/list/topic"
         topic.title = "topic"
         topic.type = ReaderListTopic.TopicType
@@ -275,32 +302,36 @@ class ReaderTopicSwiftTest: XCTestCase {
 
         // Save the new topic + posts in the contet
         var expect = expectation(description: "topics saved expectation")
-        testContextManager?.save(context!, withCompletionBlock: { () -> Void in
+        testContextManager?.save(context, withCompletionBlock: { () -> Void in
             expect.fulfill()
         })
         waitForExpectations(timeout: expectationTimeout, handler: nil)
 
         // Delete the topic and posts from the context
-        context?.delete(topic)
+        context.delete(topic)
 
         expect = expectation(description: "topics saved expectation")
-        testContextManager?.save(context!, withCompletionBlock: { () -> Void in
+        testContextManager?.save(context, withCompletionBlock: { () -> Void in
             expect.fulfill()
         })
         waitForExpectations(timeout: expectationTimeout, handler: nil)
 
         let request = NSFetchRequest<NSFetchRequestResult>(entityName: ReaderListTopic.classNameWithoutNamespaces())
-        let count = try! context?.count(for: request)
+        let count = try! context.count(for: request)
         XCTAssertTrue(count == 0, "Topic was not deleted successfully")
 
         let postRequest = NSFetchRequest<NSFetchRequestResult>(entityName: ReaderPost.classNameWithoutNamespaces())
-        let postRequestCount = try! context?.count(for: postRequest)
+        let postRequestCount = try! context.count(for: postRequest)
+        print("pistRequestCount ", postRequestCount)
         XCTAssertTrue(count == 0, "Topic posts were not deleted successfully")
     }
 
     func testTopicTitleFormatting() {
-        let context = testContextManager?.mainContext
-        let service = ReaderTopicService(managedObjectContext: context!)
+        guard let context = context else {
+            XCTFail("Context is nil")
+            return
+        }
+        let service = ReaderTopicService(managedObjectContext: context)
 
         var unformatted = "WordPress"
         var formatted = service.formatTitle(unformatted)
@@ -327,8 +358,11 @@ class ReaderTopicSwiftTest: XCTestCase {
     }
 
     func testReaderSearchTopicCreated() {
-        let context = testContextManager?.mainContext
-        let service = ReaderTopicService(managedObjectContext: context!)
+        guard let context = context else {
+            XCTFail("Context is nil")
+            return
+        }
+        let service = ReaderTopicService(managedObjectContext: context)
 
         let phrase = "coffee talk"
         let topic = service.searchTopic(forSearchPhrase: phrase)
